Added tests for gemm_strided_batched
