+++ test/unittest/blas3/blas3_gemm_batched_test.cpp
@@ -98,3 +98,44 @@ const auto BetaNonZeroLDMultipliedAlpha0 = ::testing::Combine(
     ::testing::Values(gemm_batch_type_t::strided)  // batch_type
 );
 GENERATE_GEMM_TEST(BatchGemm, BetaNonZeroLDMultipliedAlpha0);
+
+// GEMM STRIDED BATCHED tests
+template <typename scalar_t>
+const auto DefaultGemmAndGemmBatched =
+    ::testing::Combine(::testing::Values(0),              // offset
+                       ::testing::Values(1, 5),           // batch
+                       ::testing::Values(63, 128),        // m
+                       ::testing::Values(63, 128),        // n
+                       ::testing::Values(63, 128),        // k
+                       ::testing::Values('n', 't'),       // transa
+                       ::testing::Values('n', 't'),       // transb
+                       ::testing::Values<scalar_t>(3.0),  // alpha
+                       ::testing::Values<scalar_t>(7.0),  // beta
+                       ::testing::Values(1),              // lda_mul
+                       ::testing::Values(1),              // ldb_mul
+                       ::testing::Values(1),              // ldc_mul
+                       ::testing::Values(1),              // stride_a_mul
+                       ::testing::Values(1),              // stride_b_mul
+                       ::testing::Values(1)               // stride_c_mul
+    );
+GENERATE_GEMM_STRIDED_BATCHED_TEST(BatchStridedGemm, DefaultGemmAndGemmBatched);
+
+template <typename scalar_t>
+const auto AllStridedBatched =
+    ::testing::Combine(::testing::Values(0),              // offset
+                       ::testing::Values(5),              // batch
+                       ::testing::Values(128),            // m
+                       ::testing::Values(128),            // n
+                       ::testing::Values(128),            // k
+                       ::testing::Values('n', 't'),       // transa
+                       ::testing::Values('n', 't'),       // transb
+                       ::testing::Values<scalar_t>(3.0),  // alpha
+                       ::testing::Values<scalar_t>(7.0),  // beta
+                       ::testing::Values(2),              // lda_mul
+                       ::testing::Values(3),              // ldb_mul
+                       ::testing::Values(4),              // ldc_mul
+                       ::testing::Values(0, 1, 2),        // stride_a_mul
+                       ::testing::Values(0, 1, 2),        // stride_b_mul
+                       ::testing::Values(1, 2, 3)         // stride_c_mul
+    );
+GENERATE_GEMM_STRIDED_BATCHED_TEST(BatchStridedGemm, AllStridedBatched);
