+++ test/unittest/blas3/blas3_gemm_common.hpp
@@ -31,6 +31,11 @@ template <typename T>
 using gemm_arguments_t = std::tuple<int, int, int, int, int, char, char, T, T,
                                     int, int, int, gemm_batch_type_t>;
 
+template <typename T>
+using gemm_batched_strided_arguments_t =
+    std::tuple<int, int, int, int, int, char, char, T, T, int, int, int, int,
+               int, int>;
+
 // Convert batch_type=strided to interleaved on the host
 template <typename scalar_t>
 inline std::vector<scalar_t> strided_to_interleaved(
@@ -178,6 +183,106 @@ static std::string generate_name(
                      beta, ldaMul, ldbMul, ldcMul, batchType);
 }
 
+template <typename scalar_t>
+inline void verify_gemm(
+    const gemm_batched_strided_arguments_t<scalar_t> arguments) {
+  index_t offset;
+  index_t batch;
+  index_t m;
+  index_t n;
+  index_t k;
+  char transa;
+  char transb;
+  scalar_t alpha;
+  scalar_t beta;
+  index_t lda_mul;
+  index_t ldb_mul;
+  index_t ldc_mul;
+  index_t stride_a_mul;
+  index_t stride_b_mul;
+  index_t stride_c_mul;
+  std::tie(offset, batch, m, n, k, transa, transb, alpha, beta, lda_mul,
+           ldb_mul, ldc_mul, stride_a_mul, stride_b_mul, stride_c_mul) =
+      arguments;
+
+  const char ta_str[2] = {transa, '\0'};
+  const char tb_str[2] = {transb, '\0'};
+
+  auto q = make_queue();
+  blas::SB_Handle sb_handle(q);
+
+  const index_t lda = ((transa != 'n') ? k : m) * lda_mul;
+  const index_t ldb = ((transb != 'n') ? n : k) * ldb_mul;
+  const index_t ldc = m * ldc_mul;
+
+  const index_t size_a = m * k * lda_mul;
+  const index_t size_b = k * n * ldb_mul;
+  const index_t size_c = m * n * ldc_mul;
+
+  const index_t stride_a = stride_a_mul * size_a;
+  const index_t stride_b = stride_b_mul * size_b;
+  const index_t stride_c = stride_c_mul * size_c;
+
+  const index_t buffer_size_a = size_a + (batch - 1) * stride_a + offset;
+  const index_t buffer_size_b = size_b + (batch - 1) * stride_b + offset;
+  const index_t buffer_size_c = size_c + (batch - 1) * stride_c + offset;
+
+  std::vector<scalar_t> a_m(buffer_size_a);
+  std::vector<scalar_t> b_m(buffer_size_b);
+  std::vector<scalar_t> c_m_gpu(buffer_size_c);
+
+  fill_random(a_m);
+  fill_random(b_m);
+  fill_random(c_m_gpu);
+  std::vector<scalar_t> c_m_cpu = c_m_gpu;
+
+  // Use system blas to create a reference output
+  for (int i = 0; i < batch; ++i) {
+    reference_blas::gemm(ta_str, tb_str, m, n, k, alpha,
+                         a_m.data() + i * stride_a + offset, lda,
+                         b_m.data() + i * stride_b + offset, ldb, beta,
+                         c_m_cpu.data() + i * stride_c + offset, ldc);
+  }
+
+  auto m_a_gpu = blas::make_sycl_iterator_buffer<scalar_t>(buffer_size_a);
+  auto m_b_gpu = blas::make_sycl_iterator_buffer<scalar_t>(buffer_size_b);
+  auto m_c_gpu = blas::make_sycl_iterator_buffer<scalar_t>(buffer_size_c);
+
+  blas::helper::copy_to_device(sb_handle.get_queue(), a_m.data(), m_a_gpu,
+                               buffer_size_a);
+  blas::helper::copy_to_device(sb_handle.get_queue(), b_m.data(), m_b_gpu,
+                               buffer_size_b);
+  blas::helper::copy_to_device(sb_handle.get_queue(), c_m_gpu.data(), m_c_gpu,
+                               buffer_size_c);
+
+  // SYCL BLAS GEMM STRIDED BATCHED implementation
+  _gemm_strided_batched(sb_handle, transa, transb, m, n, k, alpha,
+                        m_a_gpu + offset, lda, stride_a, m_b_gpu + offset, ldb,
+                        stride_b, beta, m_c_gpu + offset, ldc, stride_c, batch);
+
+  auto event = blas::helper::copy_to_host(sb_handle.get_queue(), m_c_gpu,
+                                          c_m_gpu.data(), buffer_size_c);
+  sb_handle.wait(event);
+
+  const bool isAlmostEqual =
+      (stride_c_mul == 1)
+          ? utils::compare_vectors(c_m_gpu, c_m_cpu)
+          : utils::compare_vectors_strided(c_m_gpu, c_m_cpu, stride_c, size_c);
+  ASSERT_TRUE(isAlmostEqual);
+}
+
+template <class T>
+static std::string generate_batched_strided_name(
+    const ::testing::TestParamInfo<gemm_batched_strided_arguments_t<T>>& info) {
+  int offset, batch, m, n, k, ldaMul, ldbMul, ldcMul, stride_a_mul,
+      stride_b_mul, stride_c_mul;
+  char transa, transb;
+  T alpha, beta;
+  BLAS_GENERATE_NAME(info.param, offset, batch, m, n, k, transa, transb, alpha,
+                     beta, ldaMul, ldbMul, ldcMul, stride_a_mul, stride_b_mul,
+                     stride_c_mul);
+}
+
 /** Registers GEMM test for all supported data types
  * @param test_suite Name of the test suite
  * @param combination Combinations object
@@ -187,3 +292,14 @@ static std::string generate_name(
   BLAS_REGISTER_TEST_CUSTOM_NAME(test_suite, test_suite##combination,        \
                                  verify_gemm, gemm_arguments_t, combination, \
                                  generate_name);
+
+/** Registers GEMM Strided Batched test for all supported data types
+ * @param test_suite Name of the test suite
+ * @param combination Combinations object
+ * @see BLAS_REGISTER_TEST_CUSTOM_NAME
+ */
+#define GENERATE_GEMM_STRIDED_BATCHED_TEST(test_suite, combination)   \
+  BLAS_REGISTER_TEST_CUSTOM_NAME(test_suite, test_suite##combination, \
+                                 verify_gemm,                         \
+                                 gemm_batched_strided_arguments_t,    \
+                                 combination, generate_batched_strided_name);
